requirements.txt
watchdog==4.0.0
PyYAML==6.0.1